fix: remove extra /godot from osx/windows config path
https://github.com/godotengine/emacs-gdscript-mode/pull/138 was a good
step, but the darwin and windows paths were not quite right when
combined with the editor-settings path below - it was digging into an
extra '/godot' directory and never finding the settings. This lifts that
'godot/' dir to only the linux system type.

The docs for the paths are here:
https://docs.godotengine.org/en/stable/tutorials/io/data_paths.html#editor-data-paths

Eglot now starts automagically for me on osx! yay!